working on turnaround / subpixel corr
